requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
Pillow==11.1.0
google-generativeai==0.8.4
google-api-python-client==2.167.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3
